@@ -2,8 +2,8 @@ package edu.rosehulman.moviequotes
 
 import android.annotation.SuppressLint
 import android.content.Context
-import android.support.v7.app.AlertDialog
-import android.support.v7.widget.RecyclerView
+import androidx.appcompat.app.AlertDialog
+import androidx.recyclerview.widget.RecyclerView
 import android.util.Log
 import android.view.LayoutInflater
 import android.view.ViewGroup
