@@ -2,9 +2,9 @@ package edu.rosehulman.moviequotes
 
 
 import android.os.Bundle
-import android.support.v4.app.Fragment
-import android.support.v7.widget.LinearLayoutManager
-import android.support.v7.widget.RecyclerView
+import androidx.fragment.app.Fragment
+import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -30,7 +30,8 @@ class MovieQuoteFragment : Fragment() {
     ): View? {
         val recyclerView = inflater.inflate(R.layout.fragment_movie_quote, container, false) as RecyclerView
         adapter = MovieQuoteAdapter(context!!, uid!!)
-        recyclerView.layoutManager = LinearLayoutManager(context)
+        recyclerView.layoutManager =
+            LinearLayoutManager(context)
         recyclerView.setHasFixedSize(true)
         recyclerView.adapter = adapter
         adapter.addSnapshotListener()
